address CodeRabbit review + file-size lint (#5606)
- js_object_has_property: narrow the handle-band crash guard to the fetch/zlib
  bands (>= COMMON_HANDLE_BAND_END). Common/small handles now fall through to
  the registered small-handle property path instead of always returning false.
- headers_init_json_ptr: normalize null/undefined headers to {} in the shared
  helper so the codegen js_fetch_headers_to_json path matches the runtime thunk
  (previously serialized headers:null as "null").
- Move the headers_fetch_object_json test out of fetch/mod.rs into a sibling
  headers_json_test.rs submodule so fetch/mod.rs is back under the 2000-LOC
  CI limit (2020 -> 1992).
